% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{afr_extract}
\alias{afr_extract}
\title{Segmenting African Market From African Development Bank Group}
\usage{
afr_extract()
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Extract textual information
}

