% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{afr_list_sector}
\alias{afr_list_sector}
\title{Segmenting African Market From African Development Bank Group}
\usage{
afr_list_sector()
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
List of public sectors
}

