% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{afr_list_segment}
\alias{afr_list_segment}
\title{Segmenting African Market From African Development Bank Group}
\usage{
afr_list_segment()
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
List of available segments
}

