% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{afr_project_st}
\alias{afr_project_st}
\title{Segmenting African Market From African Development Bank Group}
\usage{
afr_project_st()
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
List of the status of projects
}

