% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sector.R
\name{afr_sct_value}
\alias{afr_sct_value}
\title{Segmenting African Market From African Development Bank Group}
\usage{
afr_sct_value(sector)
}
\description{
Segmenting African Market From African Development Bank Group
}

