% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{afr_sector_df}
\alias{afr_sector_df}
\title{Segmenting African Market From African Development Bank Group}
\usage{
afr_sector_df(sector)
}
\description{
Segmenting African Market From African Development Bank Group
}

