% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{afr_segment_df}
\alias{afr_segment_df}
\title{IE from unstructured data}
\usage{
afr_segment_df(segment, ..., pr.status = c("ongoing", "approved", "lending",
  "pipeline"), na.rm = TRUE)
}
\arguments{
\item{segment}{the name of the segment or list of segments}

\item{pr.status}{you can filter the result of segments by status.}
}
\description{
IE from unstructured data
}
\details{
Reference for to visualize all statuses \code{\link{afr_project_st}()}
Reference for to visualize all segments \code{\link{afr_list_segment}()}
}

