% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{cleanAmount}
\alias{cleanAmount}
\title{Segmenting African Market From African Development Bank Group}
\usage{
cleanAmount(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Cleaning the amount
}

