% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc.R
\name{getAmount}
\alias{getAmount}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getAmount(xx, yy)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Status of the projects are approved, ongoing, lending and pipeline
}

