% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getCountry}
\alias{getCountry}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getCountry(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Getting the name of the country
}

