% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getData}
\alias{getData}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getData(x, link)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Browsing all the subpages related to the choice of the user from the interface
}

