% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getLink}
\alias{getLink}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getLink(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Getting the link of the project
}

