% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getPageNumber}
\alias{getPageNumber}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getPageNumber(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Number of pages to explore
}

