% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getProjectDetail}
\alias{getProjectDetail}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getProjectDetail(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Getting some details on the project like implementing agency, appraisal date, approval date, start date
}

