% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getSearchReference}
\alias{getSearchReference}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getSearchReference(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Function which checks the url of all the project and their status
}

