% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getStatus}
\alias{getStatus}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getStatus(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Getting the status of the project
}

