% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{getTitle}
\alias{getTitle}
\title{Segmenting African Market From African Development Bank Group}
\usage{
getTitle(x)
}
\description{
Segmenting African Market From African Development Bank Group
}
\details{
Getting the title of the project
}

