% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{afr_sector_df}
\alias{afr_sector_df}
\title{IE from unstructured data}
\usage{
afr_sector_df(sector, ..., pr.status = c("ongoing", "approved", "lending",
  "pipeline"), na.rm = TRUE)
}
\arguments{
\item{sector}{the name of the sector or list of sectors}

\item{pr.status}{you can filter the result of sectors by status.}
}
\description{
IE from unstructured data
}
\details{
Reference for to visualize all statuses \code{\link{afr_project_st}()}
Reference for to visualize all sectors \code{\link{afr_list_sector}()}
}

